+++ CHANGELOG.md
@@ -12,8 +12,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Marked `sys::mman::{ mmap, munmap, madvise, munlock, msync }` as unsafe.
   ([#559](https://github.com/nix-rust/nix/pull/559))
-  
-<!--### Fixed-->
+
+### Fixed
+- Fixed compilation on powerpc
+  ([#569](https://github.com/nix-rust/nix/pull/569))
 
 ## [0.8.0] 2017-03-02